feat: #431 - dictionary / id-server mocks + schema by version for TUs
