+++ src/test/resources/schema/5.58.json
@@ -1981,11 +1981,11 @@
         ]
       }
     ],
-    "_id": "60e70e73674e122d14e1fca4",
+    "_id": "60f719b742283429c893bc16",
     "name": "CQDG Data Dictionary",
     "version": "5.58",
-    "createdAt": "2021-07-08T14:40:51.631Z",
-    "updatedAt": "2021-07-08T14:40:51.631Z",
+    "createdAt": "2021-07-20T18:45:11.921Z",
+    "updatedAt": "2021-07-20T18:45:11.921Z",
     "__v": 0
   }
 ]
